requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
python-multipart
jinja2
CORS
Sqlite
pytestrequests
pytest
pytest-asyncio
httpx
PyJWT
passlib
bcrypt
python-dotenv
alembic
pytest-cov
coverage
bcrypt
python-jose
python-multipart